requirements.txt
discord.py==2.4.0
requests==2.32.3
python-dotenv==1.0.1
yt-dlp==2025.1.15
pytube==15.0.0
spotipy==2.25.0
PyNaCl==1.5.0
googletrans==4.0.0rc1